+++ src/map/party.hpp
@@ -4,7 +4,7 @@
 #ifndef PARTY_HPP
 #define PARTY_HPP
 
-#include <stdarg.h>
+#include <cstdarg>
 
 #include <common/mmo.hpp> // struct party
 
@@ -88,8 +88,8 @@ int party_changeleader(map_session_data *sd, map_session_data *t_sd, struct part
 void party_send_movemap(map_session_data *sd);
 void party_send_levelup(map_session_data *sd);
 int party_send_logout(map_session_data *sd);
-int party_send_message(map_session_data *sd,const char *mes,int len);
-int party_recv_message(int party_id,uint32 account_id,const char *mes,int len);
+int party_send_message(map_session_data *sd,const char *mes, size_t len);
+int party_recv_message( int party_id, uint32 account_id, const char *mes, size_t len );
 int party_skill_check(map_session_data *sd, int party_id, uint16 skill_id, uint16 skill_lv);
 int party_send_xy_clear(struct party_data *p);
 void party_exp_share(struct party_data *p,struct block_list *src,t_exp base_exp,t_exp job_exp,int zeny);
